added CleanRL style actor, works
I added cleanrl style actor here to see if it would work just fine, as I'll need it for moving to dreamerv3 style actor loss. And it works almost fine as is, but it takes a long long time to start driving well, since there is no mechanism for exploration yet. In current main actor the exploration is forced by initializing big deviation of actions at the beginning. In dreamerv3 style update we will have the entropy term